More Menu Content
Added GameOver scene so users now can retry or quit the game to the
main menu. Added more detail to the game scene adding particles effects.
Added high score counter for end game scene as well as a leaderboard so
scores can be saved.
Increased audio volume.
Changed menu scenes so that most of the clickable objects are not
planes any more but 3d text (text mesh). Made font have its own
graphics, not just a plane color.

** Need to clean up MenuTransistion script so that it can be an
abstract class  split for menu and game scene, not both. Not really
that big of a deal but looks cleaner in the end.

** Working on touchpad keyboard popup so users can type their initials
if they got on the leaderboard.