Settings Page in place, sliders, sensitvity, audio toggle. Ball movment improved.
- Added details to the settings page. Settings page now includes audio
toggle buttons for sfx and music in game.
- Settings page includes sensitivty bar that allows users to tweak ball
movement. Mini scene appears in menu to display ball speed and movement.
- Ball movement improved. Changed Force applied to ball from
ForceMode.Acceleration to ForceMode.Impulse. As well change variable
value to a default of 9.
- Audio volume is stored in playerprefs for in game, but not saved when
app restarts. Ball sensitivity is saved for in game and app reset.
- Added game fade in for main menu. Black Screen after splash screen
seems to be a unity bug and can't be resolved no matter what.
- Cleaned up animations folder so it is more organized.

** Need to display text at end gamer if user made leaderboards

** Need to change viewport for in game. Game seems to far away from
screen, need to zoom in a little more.

** Need to clean up slider and settings controller, look like a mess.